Back after a slight delay!
Back at it again, after a break for a few days. I think CSS is simpler than other languages but get how it can become complex with super intricate websites and browsers not supporting all elements equally. I do like it and coding with it is proving straightforward, so far.