A little behind, but that's okay!
This week was a little chaotic personally, and it took me longer to finish the assignment than I had planned. However, I did eventually get through it and struggled well with some of the elements (meaning, I played with things in the 'section' element versus the 'aside' and vice versa to see the second and third order consequences. Overall, I feel like I can style a 2005 Microsoft Frontpage website - but that's okay, since the learning isn't finished!